some adjustments

+++ src/main/java/com/gizasystems/filemanagement/infrastructure/FileSystemConfig.java
@@ -17,7 +17,7 @@ public class FileSystemConfig {
 
 
 
-    private int maxFileSize;
+    private long maxFileSize;
     private String storagePath;
     private String metaStoragePath;
 }
